Moved cordovaDot into its own file renamed normalizedImagePath, and moved it into common folder. Moved historyPush into common folder. - 52 files modified
