+++ package.json
@@ -84,6 +84,7 @@
     "react-select": "~4.1.0",
     "react-share": "^4.4.0",
     "react-svg": "^11.2.5",
+    "react-text-truncate": "^0.17.0",
     "react-truncate-markup": "~5.1.0",
     "rxjs": "^6.6.7",
     "styled-components": "~5.2.1"
