+++ src/js/components/CampaignSupport/SupportButtonBeforeCompletionScreen.jsx
@@ -6,7 +6,7 @@ import styled from 'styled-components';
 import AppObservableStore from '../../stores/AppObservableStore';
 import CampaignStore from '../../stores/CampaignStore';
 import CampaignSupporterStore from '../../stores/CampaignSupporterStore';
-import historyPush from '../../utils/historyPush';
+import historyPush from '../../common/utils/historyPush';
 import { renderLog } from '../../common/utils/logging';
 import VoterStore from '../../stores/VoterStore';
