+++ src/js/components/Navigation/CampaignStartSteps.jsx
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 import { withStyles } from '@material-ui/core/styles';
 import { Done } from '@material-ui/icons';
 import CampaignStartStore from '../../stores/CampaignStartStore';
-import historyPush from '../../utils/historyPush';
+import historyPush from '../../common/utils/historyPush';
 import { renderLog } from '../../common/utils/logging';
