+++ src/js/components/Navigation/SuperSharingSteps.jsx
@@ -5,7 +5,7 @@ import { withStyles } from '@material-ui/core/styles';
 import { Done } from '@material-ui/icons';
 import { messageService } from '../../stores/AppObservableStore';
 import CampaignStore from '../../stores/CampaignStore';
-import historyPush from '../../utils/historyPush';
+import historyPush from '../../common/utils/historyPush';
 import { onStep1ClickPath, onStep2ClickPath, onStep3ClickPath, onStep4ClickPath } from '../../utils/superSharingStepPaths';
 import { renderLog } from '../../common/utils/logging';
 import VoterStore from '../../stores/VoterStore';
