+++ src/js/components/Twitter/TwitterHandleBox.jsx
@@ -2,7 +2,7 @@
 import { Button } from '@material-ui/core';
 import React, { Component } from 'react';
 import VoterStore from '../../stores/VoterStore';
-import historyPush from '../../utils/historyPush';
+import historyPush from '../../common/utils/historyPush';
 import { renderLog } from '../../common/utils/logging';
 import { extractTwitterHandleFromTextString } from '../../utils/textFormat';
 import LoadingWheel from '../../common/components/Widgets/LoadingWheel';
