+++ src/js/components/Twitter/TwitterSignIn.jsx
@@ -4,7 +4,7 @@ import { oAuthLog, renderLog } from '../../common/utils/logging';
 import $ajax from '../../utils/service';
 import Cookies from '../../common/utils/js-cookie/Cookies';
 import { isWebApp, cordovaOpenSafariView, isIOS, isAndroid } from '../../utils/cordovaUtils';
-import historyPush from '../../utils/historyPush';
+import historyPush from '../../common/utils/historyPush';
 import SplitIconButton from '../../common/components/Widgets/SplitIconButton';
 import startsWith from '../../common/utils/startsWith';
 import { shortenText } from '../../utils/textFormat';
