+++ src/js/pages/CampaignStart/CampaignStartIntro.jsx
@@ -5,7 +5,7 @@ import styled from 'styled-components';
 import { withStyles } from '@material-ui/core/styles';
 import { Button } from '@material-ui/core';
 import AppObservableStore, { messageService } from '../../stores/AppObservableStore';
-import historyPush from '../../utils/historyPush';
+import historyPush from '../../common/utils/historyPush';
 import { renderLog } from '../../common/utils/logging';
