+++ src/js/pages/CampaignSupport/CampaignSupportPayToPromote.jsx
@@ -19,7 +19,7 @@ import {
 import CampaignStore from '../../stores/CampaignStore';
 import { getCampaignXValuesFromIdentifiers, retrieveCampaignXFromIdentifiersIfNeeded } from '../../utils/campaignUtils';
 import CampaignSupportSteps from '../../components/Navigation/CampaignSupportSteps';
-import historyPush from '../../utils/historyPush';
+import historyPush from '../../common/utils/historyPush';
 import { ContentInnerWrapperDefault, ContentOuterWrapperDefault, PageWrapperDefault } from '../../components/Style/PageWrapperStyles';
 import { renderLog } from '../../common/utils/logging';
 import VoterStore from '../../stores/VoterStore';
