+++ src/js/pages/CampaignSupport/CampaignSupportShare.jsx
@@ -21,7 +21,7 @@ import CampaignShareChunk from '../../components/Campaign/CampaignShareChunk';
 import CampaignSupporterActions from '../../actions/CampaignSupporterActions';
 import CampaignSupporterStore from '../../stores/CampaignSupporterStore';
 import CampaignSupportSteps from '../../components/Navigation/CampaignSupportSteps';
-import historyPush from '../../utils/historyPush';
+import historyPush from '../../common/utils/historyPush';
 import { ContentInnerWrapperDefault, ContentOuterWrapperDefault, PageWrapperDefault } from '../../components/Style/PageWrapperStyles';
 import { renderLog } from '../../common/utils/logging';
 import SendFacebookDirectMessageButton from '../../components/Share/ShareByFacebookDirectMessageButton';
