+++ src/js/pages/PageNotFound.jsx
@@ -6,7 +6,7 @@ import { Ballot } from '@material-ui/icons';
 import { Button } from '@material-ui/core';
 import { withStyles } from '@material-ui/core/styles';
 import AppObservableStore, { messageService } from '../stores/AppObservableStore';
-import historyPush from '../utils/historyPush';
+import historyPush from '../common/utils/historyPush';
 import { renderLog } from '../common/utils/logging';
 
 class PageNotFound extends Component {
